+++ rust/cloud-storage/macro_service_urls/src/lib.rs
@@ -20,8 +20,6 @@ use url::Url;
 
 #[cfg(test)]
 mod test;
-#[cfg(test)]
-mod tests;
 
 #[cfg(test)]
 mod testing_harness {
